LayoutTests/swipe/swipe-disables-view-transition-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Skipping view transition because skipTransition() was called.
startSwipeGesture
didBeginSwipe
completeSwipeGesture
willEndSwipe
didEndSwipe
didRemoveSwipeSnapshot
